requirements.txt
Pillow==9.0.1
tensorflow==2.9.0
matplotlib==3.5.2
numpy==1.21.5
